refactor: minor PR review issues
