+++ src/components/ui/forms/personal-Info-form/personal-info-form.module.scss
@@ -4,7 +4,7 @@
 }
 
 .avatar {
-  margin-top: 14px;
+  margin-top: 26px;
 }
 
 .form {
@@ -12,5 +12,5 @@
 }
 
 .button {
-  margin-bottom: 56px;
+  margin-bottom: 60px;
 }
